#632 - rename wp view container
