+++ src/components/projects/wbs-details/work-package-page/work-package-view-container/dependencies-list/dependencies-list.module.css
