+++ src/components/projects/wbs-details/work-package-page/work-package-view-container/work-package-container-view.tsx
