+++ src/components/projects/wbs-details/work-package-page/work-package-view-container/work-package-details/work-package-details.tsx
